Rephrased the text on the personal banner
"The death toll is still counting" sounded a bit weird because the death toll isn't capable of counting, only of being counted. "Climbing", on the other hand, is a more natural alternative. I also added the number of civilian casualties to be more specific. I think "thousands" is too depersonalized and vague.