Remove the misleading claim that Symfony might stand with Ukraine
At the time of adding the link, they probably did support Ukraine. But not anymore. We noticed a number of opposite claims and silent actions in that community so keeping the mention does not represent the objective reality anymore.

There's some toxicity on their side that repeats many propagandist lines of the modern ruscist nazis. Like calling ethnic cleansing / mass murder "politics". They seem to want to distantiate themselves from the Ukrainian reality, because it's too inconvenient to know that people get killed every day.

Refs:
* https://github.com/symfony/symfony/pull/53238
* https://github.com/symfony-cli/symfony-cli/commit/7b1629faea560cc8b0c917872fae3ca62d13f250
* https://github.com/symfony-cli/symfony-cli/issues/304

Fixes #262.